add update profile Page
and minor code cleanup
