+++ components/Account/Account.js
@@ -1,16 +1,12 @@
 import { useState, useEffect } from "react";
 import { useUser, useSupabaseClient } from "@supabase/auth-helpers-react";
 import * as S from "../../pages/signup/index.styled";
-import { useRouter } from "next/router";
-import UpdateProfileForm from "../UpdateProfileForm/UpdateProfileForm";
 import LoggedInHome from "../LoggedInHome/LoggedInHome";
-import Menu from "../Menu/Menu";
-import SignUpProfile from "../../pages/signupprofile";
+import RegisterProfile from "../RegisterProfile/RegisterProfile";
 
 export default function Account({ session }) {
   const supabase = useSupabaseClient();
   const user = useUser();
-  const router = useRouter();
   const [profile, setProfile] = useState(true);
 
   useEffect(() => {
@@ -41,7 +37,7 @@ export default function Account({ session }) {
         </div>
       ) : (
         <S.signUpDiv>
-          <UpdateProfileForm session={session} />
+          <RegisterProfile session={session} />
         </S.signUpDiv>
       )}
     </div>
