+++ components/UpdateProfile/UpdateProfile.js
@@ -1,0 +1,46 @@
+import { useEffect } from "react";
+import { useUser, useSupabaseClient } from "@supabase/auth-helpers-react";
+import * as S from "../../pages/signup/index.styled";
+
+export default function UpdateProfile({ session }) {
+  const supabase = useSupabaseClient();
+  const user = useUser();
+
+  useEffect(() => {}, [session]);
+
+  // Updates the userdata!
+  async function updateProfile(event) {
+    event.preventDefault();
+    const name = event.target.name.value;
+    const username = event.target.username.value;
+    const { data, error } = await supabase
+      .from("profiles")
+      .upsert({ username: username, name: name, id: user.id });
+  }
+
+  return (
+    <S.signUpDiv>
+      <S.signUpSection>
+        <S.signUpHeading>Uppdatera din profil!</S.signUpHeading>
+        <S.signUpText>Ändra namn eller användarnamn här!</S.signUpText>
+        <S.signUpForm onSubmit={updateProfile}>
+          <S.signUpLabel htmlFor="email">Användarnamn</S.signUpLabel>
+          <S.signUpInput
+            type="text"
+            id="username"
+            name="username"
+            placeholder="Write your username here"
+          ></S.signUpInput>
+          <S.signUpLabel htmlFor="name">Namn</S.signUpLabel>
+          <S.signUpInput
+            type="text"
+            id="name"
+            name="name"
+            placeholder="Write your name here"
+          ></S.signUpInput>
+          <S.signUpButton type="submit">Skicka in</S.signUpButton>
+        </S.signUpForm>
+      </S.signUpSection>
+    </S.signUpDiv>
+  );
+}
